refactor:#33 ValidationUtil class 생성
